LayoutTests/imported/w3c/web-platform-tests/dom/events/AddEventListenerOptions-passive.any-expected.txt

CONSOLE MESSAGE: Unable to preventDefault() on event of type 'test': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'test': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'test': Inside passive event listener.

PASS Supports passive option on addEventListener only
PASS preventDefault should be ignored if-and-only-if the passive option is true
PASS returnValue should be ignored if-and-only-if the passive option is true
PASS passive behavior of one listener should be unaffected by the presence of other listeners
PASS Equivalence of option values
